// ui.frontend/src/app/core/interface.ts
export interface MappedComponentPropertiesSignal extends ReloadForceAble {
  /**
   * Path to the model associated with the current instance of the component
   */
  cqPath: InputSignal<string>;

  /**
   * Angular item name
   */
  itemName: InputSignal<string>;
}
// ...
export interface AEMContainerComponentPropertiesSignal extends MappedComponentPropertiesSignal {
  componentMapping?: typeof ComponentMapping;
  /**
   * Map of model items included in the current container
   */
  cqItems: InputSignal<{ [key: string]: Model }>;
  /**
   * Array of model item keys
   */
  cqItemsOrder: InputSignal<string[]>;
  /**
   * Class names of the current component
   */
  classNames: InputSignal<string>;
}

/**
 * Component that is allowed to be used on the page by the editor
 */
export interface AllowedComponent {
  /**
   * Path to the component under apps
   */
  path: string;

  /**
   * Title of the component
   */
  title: string;
}

/**
 * AllowedComponents collection
 */
export interface AllowedComponents {
  applicable: boolean;

  /**
   * List of allowed components
   */
  components: AllowedComponent[];
}

/**
 * Properties for the allowed components container
 */
export interface AemAllowedComponentsContainerComponentProperties extends AEMContainerComponentPropertiesSignal {
  /**
   * List of allowed components for the container
   */
  allowedComponents: InputSignal<AllowedComponents>;

  /**
   *  Label to display when there are no allowed components
   */
  _allowedComponentPlaceholderListEmptyLabel?: string;

  /**
   * Title of the placeholder list
   */
  title: InputSignal<string>;
}